Fix autoloader handling
